pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "creative-ops-llm-review-agent"
version = "0.1.0"
description = "Creative Ops LLM review showcase I built with evaluation and observability."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Tahiro" }
]
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "openai>=1.0.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "pydantic>=2.8.0",
  "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[project.scripts]
creative-ops-api = "creative_ops_llm_review_agent.api:serve"
creative-ops-eval = "creative_ops_llm_review_agent.eval_runner:main"
creative-ops-benchmark = "creative_ops_llm_review_agent.eval_runner:benchmark_main"
creative-ops-mcp = "creative_ops_llm_review_agent.mcp_server:serve"

[tool.hatch.build.targets.wheel]
packages = ["src/creative_ops_llm_review_agent"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
  "tests"
]
